+++ fish_diffusion/archs/diffsinger/diffsinger.py
@@ -1,7 +1,16 @@
+import matplotlib.pyplot as plt
+import pytorch_lightning as pl
 import torch
 import torch.nn as nn
+import wandb
+from mmengine.optim import OPTIMIZERS
+from pytorch_lightning.loggers import WandbLogger
 
 from fish_diffusion.modules.encoders import ENCODERS
+from fish_diffusion.modules.vocoders import VOCODERS
+from fish_diffusion.modules.vocoders.builder import VOCODERS
+from fish_diffusion.schedulers import LR_SCHEUDLERS
+from fish_diffusion.utils.viz import viz_synth_sample
 
 from .diffusions import DIFFUSIONS
 
@@ -108,3 +117,116 @@ class DiffSinger(nn.Module):
         output_dict["features"] = features["features"]
 
         return output_dict
+
+
+class DiffSingerLightning(pl.LightningModule):
+    def __init__(self, config):
+        super().__init__()
+        self.save_hyperparameters()
+
+        self.model = DiffSinger(config.model)
+        self.config = config
+
+        # 音频编码器, 将梅尔谱转换为音频
+        self.vocoder = VOCODERS.build(config.model.vocoder)
+        self.vocoder.freeze()
+
+    def configure_optimizers(self):
+        self.config.optimizer.params = self.parameters()
+        optimizer = OPTIMIZERS.build(self.config.optimizer)
+
+        self.config.scheduler.optimizer = optimizer
+        scheduler = LR_SCHEUDLERS.build(self.config.scheduler)
+
+        return [optimizer], dict(scheduler=scheduler, interval="step")
+
+    def _step(self, batch, batch_idx, mode):
+        assert batch["pitches"].shape[1] == batch["mel"].shape[1]
+
+        pitches = batch["pitches"].clone()
+        batch_size = batch["speaker"].shape[0]
+
+        output = self.model(
+            speakers=batch["speaker"],
+            contents=batch["contents"],
+            contents_lens=batch["contents_lens"],
+            contents_max_len=batch["contents_max_len"],
+            mel=batch["mel"],
+            mel_lens=batch["mel_lens"],
+            mel_max_len=batch["mel_max_len"],
+            pitches=batch["pitches"],
+            pitch_shift=batch.get("key_shift", None),
+        )
+
+        self.log(f"{mode}_loss", output["loss"], batch_size=batch_size, sync_dist=True)
+
+        if mode != "valid":
+            return output["loss"]
+
+        x = self.model.diffusion(output["features"])
+
+        for idx, (gt_mel, gt_pitch, predict_mel, predict_mel_len) in enumerate(
+            zip(batch["mel"], pitches, x, batch["mel_lens"])
+        ):
+            image_mels, wav_reconstruction, wav_prediction = viz_synth_sample(
+                gt_mel=gt_mel,
+                gt_pitch=gt_pitch[:, 0],
+                predict_mel=predict_mel,
+                predict_mel_len=predict_mel_len,
+                vocoder=self.vocoder,
+                return_image=False,
+            )
+
+            wav_reconstruction = wav_reconstruction.to(torch.float32).cpu().numpy()
+            wav_prediction = wav_prediction.to(torch.float32).cpu().numpy()
+
+            # WanDB logger
+            if isinstance(self.logger, WandbLogger):
+                self.logger.experiment.log(
+                    {
+                        f"reconstruction_mel": wandb.Image(image_mels, caption="mels"),
+                        f"wavs": [
+                            wandb.Audio(
+                                wav_reconstruction,
+                                sample_rate=44100,
+                                caption=f"reconstruction (gt)",
+                            ),
+                            wandb.Audio(
+                                wav_prediction,
+                                sample_rate=44100,
+                                caption=f"prediction",
+                            ),
+                        ],
+                    },
+                )
+
+            # TensorBoard logger
+            if isinstance(self.logger, TensorBoardLogger):
+                self.logger.experiment.add_figure(
+                    f"sample-{idx}/mels",
+                    image_mels,
+                    global_step=self.global_step,
+                )
+                self.logger.experiment.add_audio(
+                    f"sample-{idx}/wavs/gt",
+                    wav_reconstruction,
+                    self.global_step,
+                    sample_rate=44100,
+                )
+                self.logger.experiment.add_audio(
+                    f"sample-{idx}/wavs/prediction",
+                    wav_prediction,
+                    self.global_step,
+                    sample_rate=44100,
+                )
+
+            if isinstance(image_mels, plt.Figure):
+                plt.close(image_mels)
+
+        return output["loss"]
+
+    def training_step(self, batch, batch_idx):
+        return self._step(batch, batch_idx, mode="train")
+
+    def validation_step(self, batch, batch_idx):
+        return self._step(batch, batch_idx, mode="valid")
