+++ fish_diffusion/utils/inference.py
@@ -1,9 +1,9 @@
 import torch
 
-from train import FishDiffusion
+from fish_diffusion.archs.diffsinger.diffsinger import DiffSingerLightning
 
 
-def load_checkpoint(config, checkpoint, device="cuda") -> FishDiffusion:
+def load_checkpoint(config, checkpoint, device="cuda", model_cls=DiffSingerLightning):
     """Load checkpoint from path
 
     Args:
@@ -12,10 +12,10 @@ def load_checkpoint(config, checkpoint, device="cuda") -> FishDiffusion:
         device: device
 
     Returns:
-        FishDiffusion: model
+        model
     """
 
-    model = FishDiffusion(config)
+    model = model_cls(config)
     state_dict = torch.load(checkpoint, map_location="cpu")
 
     if "state_dict" in state_dict:  # Checkpoint is saved by pl
